Convert string variables to lowercase
It's assumed that the binary yes/no variables are lowercase, however if
the user enters them in uppercase the string comparisons will fail. The
declare -l ensures that whatever is written in auter.conf is converted
to lowercase for the comparisons to work correctly.

Also declare the MAXDELAY as an integer to make operations later more
reliable.