Export config options for use in pre/post scripts (#106)
* Export config options for use in pre/post scripts

Currently pre/post scripts run via run_script do not inherit the config
from auter - there is no way to know what CONFIGSET/CONFIGFILE are
without sourcing the default config file. If auter is being
run with --config <file> then there is no way to know what options are
set without passing all of them as arguments.

Exporting the config options removes this issue.

* Remove declare from CONFIGFILE and swap unset for "" on CONFIGSET

Rather than unset CONFIGSET and remove it from the shell, set it to "".
Remove unnecessary declare from CONFIGFILE as it's already exported at
the start.

Some additional formatting clean up on the other options.

* Convert string variables to lowercase

It's assumed that the binary yes/no variables are lowercase, however if
the user enters them in uppercase the string comparisons will fail. The
declare -l ensures that whatever is written in auter.conf is converted
to lowercase for the comparisons to work correctly.

Also declare the MAXDELAY as an integer to make operations later more
reliable.