Fix auto-reboot not working (#143)
AUTOREBOOT="yes" is not working due to the REBOOTCALL variable being
declared in the apply_update function.

This moves the REBOOTCALL declaration to the start of the main auter
script and removes the declare from within the function. We maintain the
default no auto-reboot, however the global variable is now updated
correctly.

Ignore the shellcheck warning for an unused REBOOTCALL as it has been
exported from the main auter script which shellcheck is unaware of.