V0.12 (#193)
* Changed compare to the develop branch. (#125)

* Changed compare to the develop branch.

* Adjusted the local branch name in .travis.yml

* Update .travis.yml

* Moved the prep rpm cleanup. Also added an rpm cleanup after apply has run (#138)

*  Fix auto-reboot not working (#143) (#145)

AUTOREBOOT="yes" is not working due to the REBOOTCALL variable being
declared in the apply_update function.

This moves the REBOOTCALL declaration to the start of the main auter
script and removes the declare from within the function. We maintain the
default no auto-reboot, however the global variable is now updated
correctly.

Ignore the shellcheck warning for an unused REBOOTCALL as it has been
exported from the main auter script which shellcheck is unaware of.

* Remove output files greater than the ROTATE value (#146)

* Remove output files greater than the ROTATE value

If the ROTATE value is reduce then we need to clean up the old files
with indexes greater than ROTATE.

* Swap local with declare for consistency

* Log auter prep even when there are no updates available (#139)

This has been tested and is working as expected

* Update prepare_updates to always log to output

Always capture PREPOUTPUT, regardless of the outcome, making it easier
to debug no updates during prep time and subsequently apply time.
Include the date in the PREPOUTPUT, again for easier debugging.

Also declare local function variables at the top of the function for
clarity.

* Swap local for declare for consistency

* Improved Makefile for rpm build sources (#131)

* Improved Makefile for rpm build sources

* Moved the lintian line to a better place

* Adjusted auter status message when run as non-root user (#149)

* Suppress shutdown wall messages if the option is valid (#147)

* Suppress shutdown wall messages if the option is valid

* Added STDOUT and STDERR redirection to the shutdown command

* Improved if condition for reboot and captured shutdown output

* Removed attempt to capture reboot output

* Contrib conditional reboot (#152)

* Initial commit for conditional reboot

* Shellcheck fixes

* added Oracle Linux as distribution

Added the Oracle Linux distribution

* Reformatted output

Reformatted the output so one package is on one line. When there are
multiple matches the logit-line can become really long.

* Corrected issues from #151 and fixed shellcheck problems (#153)

* Corrected issues from #151 and fixed shellcheck problems

* Fixed the DOWNLOADOPTION array to have 2 elements. Converted DEBS variable to array

* Man page maintenance (#150)

* Improved the help2man command in the spec file

* Added auter.conf man page and also improved auter man page

* Added missing infor to auter.help2man-sections

* Added missing auter.conf.man to the %files list

* Fix AUTOREBOOT!! (#155)

Only change needed as auter has already been patched during the yum/dnf
remediation.

* Documentation update (#156)

* Added build steps for creating debs and rpms. Also renamed buildGuide.txt to buildGuide.md for better readability

* Changes made: 1) Improved the Make file definition for Distribution2) Fixed typo for line 62 3) Corrected command in line 96

* Schedule the reboot and update PATH for lsof (#158)

* Schedule the reboot and update PATH for lsof

Schedule the auter --reboot to take place one auter has finished
updating and update PATH to include /usr/sbin and /usr/local/sbin for
lsof run by crontab.

* Update sleep time and reboot message

Check on the auter PIDFILE every 5 seconds rather than 1 second and
alter the reboot message so that it no longer implies that the reboot
has been scheduled for a specific time.

* Rename conditionalReboot.sh to 99-conditionalReboot.sh

The script should be executed last hence the 99.

* Feature/update script documentation (#161)

* Update man page, add section on hook execution order

Update section headers and re-order a couple of the sections. Also
include a new section on the script execution order and make a
recommendation to use padded leading numbers in the filenames to ensure
correct order.

* Remove auter.help2man as it has been replace

Replaced by auter.help2man-sections.

* Update execution order section with note

Note about the case-insensitive sorting that bash path expansion is
doing.

* Correct grammar and add section to contrib README

Updates with recommendations made in the PR.

* Removed unused shellcheck excludes and commented the used excludes (#164)

* Initial commit of configsnap scripts (#148)

* Initial commit of configsnap scripts

* Fixed shellcheck issues

* Removed debian based OS's from the README.md

* Contrib readme update (#165)

* Updated contrib README.md

* Updated configsnap README.md

* Updated conditionalReboot README and minor code adjustment

* Various Skip script options (#160)

* Added --skip-all-scripts option

* Added --skip-phase-scripts option

* Added --skip-scripts-by-name option

* Fixed up the -h output (Also used to generate man page)

* Fixes for review comments

* Travis debbuild (#119)


    Rename the .deb to include the distribution
    Checks need to be done to confirm the debs are created.
    Fail flag needs to be set for each build

* Remove needless use of %defattr (#169)

Resolves #167 
For reference:
https://src.fedoraproject.org/rpms/auter/c/f68da342d48235b0b5ed0c4bd120d9f9d870555c?branch=master

* Contrib conditional reboot kernel check (#168)

Resolves #166

* lsof fix for RHEL/CentOS 6 (#172)

Resolves #171 
1) Added lsof as a dependency in README.md.
2) Added a check that lsof executable exists

* Version bump and changelog update

* Version bump in spec file

* REBOOTCALL is now 0 by default so the test is failing (#182)

* REBOOTCALL is now 0 by default so the test is failing

Swap the string check for an regex check against the minimum required
arguments. When none of the required args are present, run print_help and exit.

* Match against the full long argument including leading dashes

* Added config to create man page for auter.conf. Also adjusted that man page to be a man5 page not man8 (as per man man) (#183)

* adding PATH test in main auter script and setting noninteractive apt mode (#191)

* adding PATH test in main auter script and setting DEBIAN_FRONTED=noninteractive in apt module

* improve PATH testing on PR 191

* Updated manual tests (#184)

* improved testing template in HACKING.md

* improved testing template layout in HACKING.md
Adjusted the guideline before the template
Corrected some grammar errors

* Added instructions to downgrade package in .deb based distros

* Added a check for man auter.conf

* minor fixes as per @LukeHandle